+++ src/corner_kick/.prettierrc
@@ -1,8 +1,8 @@
 {
-  "printWidth": 90,
-  "semi": true,
-  "singleQuote": true,
-  "tabWidth": 2,
-  "arrowParens": "always",
-  "trailingComma": "all"
+    "printWidth": 90,
+    "semi": true,
+    "singleQuote": true,
+    "tabWidth": 4,
+    "arrowParens": "always",
+    "trailingComma": "all"
 }
