+++ src/corner_kick/README.md
@@ -4,7 +4,7 @@ Corner Kick is an Electron app that allows for the visualization of UBC Thunderb
 
 ## How to use
 
-After running `catkin_make`, you can start the visualizer by running:
+You can start the visualizer by running:
 
 ```
 roslaunch corner_kick corner_kick.launch
@@ -41,24 +41,23 @@ A window should open with the application in development mode.
 
 The project is defined as follow:
 
-- config - contains webpack (build tool) configuration files
-- launch - contains roslaunch file
-- public - contains the HTML base file and the Electron start code
-- src
-  - components
-    - containers - contains React stateful components
-    - ui - contains repeatable UI components
-  - modules - contains high-level UI components (logger, visualizer, etc.)
-  - service - contains various communication layers (including the service communicating with ROS)
-  - types - contains ROS message type schema
-  - App.tsx - top-level React component
-  - index.css - base css styling
-  - index.tsx - initializes React
-- .prettierrc - configuration for autoformating tool
-- CMakeLists.txt - cmake build script
-- module.d.ts - definition file for non-typescript libraries
-- package.json - specifies dependencies, dev and build scripts.
-- package.xml - specifies ros dependencies
-- tsconfig.json - configuration file for typescript compiler
-- tslint.json - configuration for typescript linter
-- yarn.lock - autogenerated lock file for javascript dependencies
+-   config - contains webpack (build tool) configuration files
+-   launch - contains roslaunch file
+-   src
+    -   main - contains code for the main process
+    -   renderer - contains code for the rendering process
+        -   components
+            -   containers - contains React stateful components
+            -   ui - contains repeatable UI components
+        -   modules - contains high-level UI components (logger, visualizer, etc.)
+    -   shared - contains code shared by the two process
+        -   ros - contains the service communicating with ROS
+            -   types - contains ROS message type schema
+-   .prettierrc - configuration for autoformating tool
+-   CMakeLists.txt - cmake build script
+-   module.d.ts - definition file for non-typescript libraries
+-   package.json - specifies dependencies, dev and build scripts.
+-   package.xml - specifies ros dependencies
+-   tsconfig.json - configuration file for typescript compiler
+-   tslint.json - configuration for typescript linter
+-   yarn.lock - autogenerated lock file for javascript dependencies
