+++ src/corner_kick/config/webpack.config.prod.js
@@ -1,7 +1,21 @@
+/**
+ * @fileoverview Specifies the webpack configuration for production.
+ * Extends the general webpack configuration.
+ *
+ * @see {@link ./webpack.config.js}
+ */
+
 const webpackConfig = require('./webpack.config.js');
 
-// The production webpack config. Uses settings defined in ./webpack.config.js.
-module.exports = {
-  ...webpackConfig,
-  mode: 'production',
-};
+module.exports = [
+    // Rendering build configuration
+    {
+        ...webpackConfig[0],
+        mode: 'production',
+    },
+    // Main build configuration
+    {
+        ...webpackConfig[1],
+        mode: 'production',
+    },
+];
