+++ src/corner_kick/module.d.ts
@@ -1,8 +1,6 @@
 /**
- * This definition file defines third-party libraries that
+ * @fileoverview This definition file defines third-party libraries that
  * do not include any typescript definitions.
- *
- * TODO create typescript definition files for these libraries.
  */
 declare module 'roslib';
 declare module 'react-panelgroup';
