+++ src/corner_kick/package.json
@@ -1,56 +1,59 @@
 {
-  "name": "corner-kick",
-  "description": "A visualization tool for UBC Thunderbots",
-  "main": "public/main.js",
-  "author": {
-    "name": "UBC Thunderbots",
-    "email": "ubcrobocup@gmail.com"
-  },
-  "homepage": "./",
-  "version": "1.0.0",
-  "private": true,
-  "scripts": {
-    "start": "concurrently \"yarn webpack-dev-server --config ./config/webpack.config.dev.js\" \"wait-on http://localhost:8080 && electron .\"",
-    "build": "webpack --config ./config/webpack.config.prod.js",
-    "prod": "ELECTRON_ENV=production electron ."
-  },
-  "dependencies": {
-    "@blueprintjs/core": "3.7.0",
-    "@types/pixi.js": "4.8.2",
-    "electron-is-dev": "1.0.1",
-    "electron-json-storage": "4.1.4",
-    "lodash": "4.17.11",
-    "pixi-viewport": "3.9.0",
-    "pixi.js": "4.8.2",
-    "react": "16.6.0",
-    "react-dom": "16.6.0",
-    "react-panelgroup": "1.0.7",
-    "react-pixi-fiber": "0.5.0",
-    "roslib": "0.20.0",
-    "styled-components": "4.0.2",
-    "unstated": "2.1.1"
-  },
-  "devDependencies": {
-    "@types/electron-json-storage": "4.0.0",
-    "@types/lodash": "4.14.117",
-    "@types/react": "16.4.18",
-    "@types/react-dom": "16.0.9",
-    "@types/styled-components": "4.0.3",
-    "concurrently": "4.0.1",
-    "css-loader": "1.0.0",
-    "electron": "3.0.6",
-    "electron-builder": "20.28.4",
-    "html-loader": "0.5.5",
-    "html-webpack-plugin": "3.2.0",
-    "prettier": "1.14.3",
-    "style-loader": "0.23.1",
-    "ts-loader": "5.2.2",
-    "tslint": "5.11.0",
-    "tslint-react": "3.6.0",
-    "typescript": "3.1.3",
-    "wait-on": "3.1.0",
-    "webpack": "4.23.1",
-    "webpack-cli": "3.1.2",
-    "webpack-dev-server": "3.1.10"
-  }
+    "name": "corner-kick",
+    "description": "A visualization tool for UBC Thunderbots",
+    "main": "build/main.js",
+    "author": {
+        "name": "UBC Thunderbots",
+        "email": "ubcrobocup@gmail.com"
+    },
+    "homepage": "./",
+    "version": "1.0.0",
+    "private": true,
+    "scripts": {
+        "prestart": "webpack --config ./config/webpack.config.main.dev.js",
+        "start": "concurrently \"yarn start:server\" \"yarn start:electron\"",
+        "start:server": "yarn webpack-dev-server --config ./config/webpack.config.renderer.dev.js",
+        "start:electron": "wait-on http://localhost:8080 && electron .",
+        "build": "webpack --config ./config/webpack.config.prod.js",
+        "prod": "ELECTRON_ENV=production electron ."
+    },
+    "dependencies": {
+        "@blueprintjs/core": "3.7.0",
+        "@types/pixi.js": "4.8.2",
+        "electron-is-dev": "1.0.1",
+        "electron-json-storage": "4.1.4",
+        "lodash": "4.17.11",
+        "pixi-viewport": "3.9.0",
+        "pixi.js": "4.8.2",
+        "react": "16.6.0",
+        "react-dom": "16.6.0",
+        "react-panelgroup": "1.0.7",
+        "react-pixi-fiber": "0.5.0",
+        "roslib": "0.20.0",
+        "styled-components": "4.0.2",
+        "unstated": "2.1.1"
+    },
+    "devDependencies": {
+        "@types/electron-json-storage": "4.0.0",
+        "@types/lodash": "4.14.117",
+        "@types/react": "16.4.18",
+        "@types/react-dom": "16.0.9",
+        "@types/styled-components": "4.0.3",
+        "concurrently": "4.0.1",
+        "css-loader": "1.0.0",
+        "electron": "3.0.6",
+        "electron-builder": "20.28.4",
+        "html-loader": "0.5.5",
+        "html-webpack-plugin": "3.2.0",
+        "prettier": "1.14.3",
+        "style-loader": "0.23.1",
+        "ts-loader": "5.2.2",
+        "tslint": "5.11.0",
+        "tslint-react": "3.6.0",
+        "typescript": "3.1.3",
+        "wait-on": "3.1.0",
+        "webpack": "4.23.1",
+        "webpack-cli": "3.1.2",
+        "webpack-dev-server": "3.1.10"
+    }
 }
