+++ src/corner_kick/scripts/start_corner_kick.sh
@@ -1,4 +1,10 @@
 #!/bin/bash
+# Start script for Corner Kick.
+
+# Get the current directory of the script.
 DIRECTORY=`dirname $0`
+
+# Move to this directory and start the production build of
+# Corner Kick.
 cd $DIRECTORY
 yarn prod
