--- src/corner_kick/src/modules/Visualizer/index.tsx
@@ -1,1 +1,0 @@
-export { Canvas as Visualizer } from './Canvas';
