+++ src/corner_kick/src/renderer/index.css
@@ -1,0 +1,28 @@
+/** 
+ * Provides some styling to remove base styling provided
+ * by Electron
+ */
+
+body {
+    /* Body should take 100% of the window */
+    width: 100vw;
+    height: 100vh;
+    margin: 0;
+    padding: 0;
+    font-family: sans-serif;
+    overflow: hidden;
+}
+
+#root {
+    /* root should take 100% of body */
+    width: 100%;
+    height: 100%;
+}
+
+canvas {
+    /* 
+    Canvas should take 100% of its parent.
+    Necessary for automatic canvas resizing.
+  */
+    width: 100% !important;
+}
