+++ src/corner_kick/src/renderer/index.html
@@ -1,3 +1,8 @@
+<!-- 
+  The HTML that loads when our application window opens.
+  Webpack will inject our application bundle in this file.
+-->
+
 <!DOCTYPE html>
 <html lang="en">
