+++ src/corner_kick/src/renderer/modules/Visualizer/index.tsx
@@ -1,0 +1,4 @@
+/**
+ * @fileoverview Exposes Visualizer interfaces to other modules.
+ */
+export { Canvas as Visualizer } from './Canvas';
