--- src/corner_kick/src/types/standardTopics.ts
@@ -1,17 +1,0 @@
-/**
- * Definition file for standard ROS topics. This includes message type that are
- * standard to ROS.
- */
-
-/**
- * ROS topic used for logging purposes
- */
-export interface IROSOut {
-  file: string;
-  function: string;
-  level: number;
-  line: number;
-  msg: string;
-  name: string;
-  topics: string[];
-}
